app/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-filter==23.3
django-mqtt==0.4.0
djangorestframework==3.14.0
Markdown==3.5
paho-mqtt==1.6.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
